Feature + Clean + Potion Management
- General: Potion option is now available in Buff UI
- General: Change default settings to None for Potion
- General: Change default settings to Never for Interrupt
- DRG: Add Potion in rotation + Settings
- DRG: Add ForceLimitBreak
