+++ Magitek/Logic/Bard/Utility.cs
@@ -1,7 +1,6 @@
 ﻿using ff14bot;
 using ff14bot.Managers;
 using Magitek.Extensions;
-using Magitek.Logic.Roles;
 using Magitek.Models.Bard;
 using Magitek.Models.Scholar;
 using Magitek.Toggles;
@@ -14,13 +13,6 @@ namespace Magitek.Logic.Bard
 {
     internal static class Utility
     {
-        public static async Task<bool> UsePotion()
-        {
-            if (Spells.RagingStrikes.IsKnown() && !Spells.RagingStrikes.IsReady(4000))
-                return false;
-
-            return await PhysicalDps.UsePotion(BardSettings.Instance);
-        }
 
         public static async Task<bool> Troubadour()
         {
