+++ Magitek/Logic/Dancer/Buff.cs
@@ -4,6 +4,7 @@ using ff14bot.Managers;
 using ff14bot.Objects;
 using Magitek.Enumerations;
 using Magitek.Extensions;
+using Magitek.Logic.Roles;
 using Magitek.Models.Dancer;
 using Magitek.Utilities;
 using System;
@@ -248,6 +249,14 @@ namespace Magitek.Logic.Dancer
             }
 
             return c.CurrentJob == ClassJobType.Adventurer ? 70 : 0;
+        }
+
+        public static async Task<bool> UsePotion()
+        {
+            if (Spells.Devilment.IsKnown() && !Spells.Devilment.IsReady(11000))
+                return false;
+
+            return await PhysicalDps.UsePotion(DancerSettings.Instance);
         }
 
     }
