+++ Magitek/Logic/Samurai/Buff.cs
@@ -1,6 +1,7 @@
 ﻿using ff14bot;
 using ff14bot.Managers;
 using Magitek.Extensions;
+using Magitek.Logic.Roles;
 using Magitek.Models.Samurai;
 using Magitek.Utilities;
 using System.Linq;
@@ -106,6 +107,14 @@ namespace Magitek.Logic.Samurai
             SamuraiRoutine.InitializeFillerVar(false, false); // Remove Filler after Even Minutes Burst
 
             return true;
+        }
+
+        public static async Task<bool> UsePotion()
+        {
+            if (Spells.HissatsuSenei.IsKnown() && !Spells.HissatsuSenei.IsReady(4000))
+                return false;
+
+            return await PhysicalDps.UsePotion(SamuraiSettings.Instance);
         }
     }
 }
