+++ Magitek/Logic/Samurai/Utility.cs
@@ -2,7 +2,6 @@
 using ff14bot.Managers;
 using Magitek.Enumerations;
 using Magitek.Extensions;
-using Magitek.Logic.Roles;
 using Magitek.Models.Account;
 using Magitek.Models.Samurai;
 using Magitek.Toggles;
@@ -15,13 +14,6 @@ namespace Magitek.Logic.Samurai
 {
     internal static class Utility
     {
-        public static async Task<bool> UsePotion()
-        {
-            if (Spells.HissatsuSenei.IsKnown() && !Spells.HissatsuSenei.IsReady(4000))
-                return false;
-
-            return await PhysicalDps.UsePotion(SamuraiSettings.Instance);
-        }
 
         public static async Task<bool> Hagakure()
         {
@@ -84,7 +76,6 @@ namespace Magitek.Logic.Samurai
             }
 
             return false;
-
         }
 
         public static async Task<bool> TrueNorth()
