+++ Magitek/Rotations/Machinist.cs
@@ -102,7 +102,7 @@ namespace Magitek.Rotations
                     //Utility
                     if (await PhysicalDps.ArmsLength(MachinistSettings.Instance)) return true;
                     if (await PhysicalDps.Interrupt(MachinistSettings.Instance)) return true;
-                    if (await Utility.UsePotion()) return true;
+                    if (await Cooldowns.UsePotion()) return true;
 
                     //Pets
                     if (await Pet.RookQueen()) return true;
@@ -127,7 +127,7 @@ namespace Magitek.Rotations
                     if (await PhysicalDps.ArmsLength(MachinistSettings.Instance)) return true;
                     if (await PhysicalDps.SecondWind(MachinistSettings.Instance)) return true;
                     if (await PhysicalDps.Interrupt(MachinistSettings.Instance)) return true;
-                    if (await Utility.UsePotion()) return true;
+                    if (await Cooldowns.UsePotion()) return true;
 
                     //Cooldowns
                     if (await Cooldowns.Reassemble()) return true;
